Teste das exceptions foram retirados de ProductServiceTest e colocados em uma classe separada ProductServiceExceptionTest
